requirements.txt
Django==3.2
Gunicorn==20.1.0
psycopg2==2.9.3
djangorestframework==3.13.1